+++ Assets/_GameAssets/Scripts/Helpers/Others.cs
@@ -8,4 +8,8 @@ public class Others
     public const String MENU_SCENE = "MenuScene";
     public const String GROUND_LAYER = "Ground";
     public const String FLOOR_LAYER = "Floor";
+    public const float CAMERA_SHAKE_AMP_FOR_DAMAGE = 1f;
+    public const float CAMERA_SHAKE_AMP_FOR_INSTA_DEAD = 1f;
+    public const float CAMERA_SHAKE_DURATION_FOR_DAMAGE = 1f;
+    public const float CAMERA_SHAKE_DURATION_FOR_INSTA_DEAD = 2f;
 }